+++ HRIS.API/Models/Dto/ActiveStaffDto.cs
@@ -8,7 +8,7 @@ namespace HRIS.API
     public class ActiveStaffDto
     {
         public int Total { get; set; }
-        public int RowNum { get; set; }
+        public Int64 RowNum { get; set; }
         public string EIN { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
